Add openssl to nix flake shell
Expose openssl so that we can use poetry install to build psycopg2

On the other hand, since poetry2nix is building the environment for us,
it feels like we are doing twice the work here.

It would therefore be great if we can just expose more things that
poetry2nix is building for us inside the nix develop shell.