BE: Fix flaky user email test
Previously, we created the user email confirm token a second time and
check if it is part of the email. If user_make_token, and email
rendering do not happen within the same second, the token will be
different. CircleCI pointed out that this test was flaky.

I have done the same for the user password reset email.