BE: Conditionally ignore Redis TLS urls (yikes)
When we run the Projectify backend locally, we don't always have a TLS
cert at hand.

The way the backend uses TLS for Redis right now is not very smart
anyway, since we don't validate certificates at all.

The better way to handle this in the future is to either forego TLS
at all and only use Redis in a trusted network OR use proper TLS
certificates with Heroku's Redis, which it doesn't right now. Hence
the original workaround of disabled cert validation.