FE: Give cosmetic divs role=presentation
Two <divs> that are part of the overall flex layout are given
role=presentation to ensure that they are not part of the accessibility
tree.

The accessibility tree was reviewed using Firefox v129's accessibility
inspector. After changing the role to "presentation", the two divs
modified in this do not show up anymore.