BE: Fix email urls for non-django frontend
I changed the emails in a way that wouldn't work with the old frontend.
When the URLs that we resolve insidde the emails don't exist, emails
can't be sent out correctly. Because the new frontend is disabled in the
production version of Projectify, the URLs didn't resolve correctly.

This patch adds appropriate if/else branches to add the correct URLs into the
context. This patch also fixes the email bodies to use `{{ url }}`
instead of resolving the URL inside the template.